Test_task/Rocketdata/requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
django-cors-headers==3.11.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Faker==11.3.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
postgres==4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
